pyproject.toml
[project]
name = "asr-engine"
version = "0.1.0"
description = "本地语音识别服务"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "jinja2>=3.1.0",
    "zhconv>=1.4.2",
    "tomli>=2.0.0; python_version < '3.11'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
